# Backyard-app/openapi.yaml
openapi: 3.0.1
info:
  title: Backyard
  description: This is a Backyard server.
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: pietrusjakub@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://backyard.swagger.io/v1'
  - url: 'http://backyard.swagger.io/v1'
...
paths:
...
  /role:
    get:
      tags:
        - role
      summary: Get roles list
      operationId: getRolesList
      parameters:
        - name: sort
          in: query
          description: How to sort the results
          schema:
            type: string
            default: id
            enum:
              - id
              - name
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    put:
      tags:
        - role
      summary: Update an existing role
      operationId: updateRole
      requestBody:
        description: Role object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
          application/xml:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Role not found
          content: {}
        '405':
          description: Validation exception
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - role
      summary: Create role
      operationId: createRole
      requestBody:
        description: Created role object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  '/role/{roleId}':
    delete:
      tags:
        - role
      summary: Deletes a role
      operationId: deleteRole
      parameters:
        - name: roleId
          in: path
          description: Game id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
          content: {}
        '404':
          description: Game not found
          content: {}
  '/role/user/{username}':
    get:
      tags:
        - role
      summary: Get user roles
      operationId: getUserRoles
      parameters:
        - name: username
          in: path
          description: name that of user we want to get roles
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '400':
          description: Invalid username supplied
          content: {}
        '404':
          description: User not found
          content: {}
...
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: Role
...